+++ src/main/java/org/example/internet_shop/dto/CartItemDto.java
@@ -3,14 +3,16 @@ package org.example.internet_shop.dto;
 import lombok.Data;
 import org.example.internet_shop.Entity.CartItem;
 
+import java.math.BigDecimal;
+
 @Data
 public class CartItemDto {
     private Long id;
     private Long productId;
     private String productName;
-    private double productPrice;
+    private BigDecimal productPrice;
     private Integer quantity;
-    private double totalPrice;
+    private BigDecimal totalPrice;
     private String previewImageUrl;
 
     public CartItemDto(CartItem cartItem) {
